% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsrAssign.R
\name{dsrAssign}
\alias{dsrAssign}
\title{Load data in the remote session(s)}
\usage{
dsrAssign(
  symbol,
  table,
  db_connection,
  collist = NULL,
  where_clause = NULL,
  row_limit = NULL,
  async = TRUE,
  datasources = NULL
)
}
\arguments{
\item{symbol}{a character, the name of the data frame}

\item{table}{a character, the name of a table in the remote database}

\item{db_connection}{a character, the name of the SQLFlexClient object. This object has to exist already in the remote
session(s) and is created with the function datashield.assign.resource.}

\item{collist}{a vector , names of the columns to load. If NULL (the default) all the columns will be loaded.}

\item{where_clause}{a character, an optional where clause to the sql statement that will load the table. It must not begin with "where".}

\item{async}{same as in datashield.assign}

\item{datasources}{same as in datashield.assign}
}
\description{
Assign data to a data frame from a resource of the type SQL Query
}
