% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsrEstimateObjectSize.R
\name{dsrEstimateObjectSize}
\alias{dsrEstimateObjectSize}
\title{Virtual object size}
\usage{
dsrEstimateObjectSize(
  virtual_objects,
  db_connection,
  async = TRUE,
  datasources = NULL
)
}
\arguments{
\item{virtual_objects}{a character vector containing either database view names or valid queries}

\item{db_connection}{a character, the name of the SQLFlexClient object . This object has to exist already in the remote
session(s) and is created with the function datashield.assign.resource}

\item{async}{same as in datashield.assign}

\item{datasources}{same as in datashi eld.assign}
}
\value{
a data frame containing the column names and data types
}
\description{
Estimate the size in memory of views or query result sets
}
