% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsrSetDbParam.R
\name{dsrSetDbParam}
\alias{dsrSetDbParam}
\title{Set parameters in the remote databases}
\usage{
dsrSetDbParam(db_connection, param_list, async = TRUE, datasources = NULL)
}
\arguments{
\item{db_connection}{a character, the name of the SQLFlexClient object. This object has to exist already in the remote
session(s) and is created with the function datashield.assign.resource.}

\item{param_list}{a list of the type (parameter_name = parameter_value)}

\item{async}{same as in datashield.assign}

\item{datasources}{same as in datashield.assign}
}
\description{
This function requires advanced knowledge of the database systems inolved, it should be used
only under supervision from a database administrator.
}
