% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsShowColumns.R
\name{dsrShowColumns}
\alias{dsrShowColumns}
\title{Describe a table}
\usage{
dsrShowColumns(
  db_connection,
  table_name,
  schema_name = "public",
  async = TRUE,
  datasources = NULL
)
}
\arguments{
\item{db_connection}{a character, the name of the SQL Query. This object has to exist already in the remote
session(s) and is created with the function datashield.assign.resource}

\item{table_name}{a character, the name of the table}

\item{schema_name}{a caracter, the name of the schema in which the table resides.}

\item{async}{same as in datashield.assign}

\item{datasources}{same as in datashield.assign}
}
\value{
a data frame containing the column names and data types
}
\description{
List the columns and data types of a table from a resource of the type SQL Query
}
