% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsrShowTables.R
\name{dsrShowTables}
\alias{dsrShowTables}
\title{List available tables}
\usage{
dsrShowTables(db_connection, async = TRUE, datasources = NULL)
}
\arguments{
\item{db_connection}{a character, the name of the SQL Query. This object has to exist already in the remote
session(s) and is created with the function datashield.assign.resource.}

\item{async}{same as in datashield.assign}

\item{datasources}{same as in datashield.assign}
}
\value{
a data frame containing the available tables and their respective schemas
}
\description{
Retrieve the list of tables from  a resource of the type SQL Query
}
