% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsrAssign.R
\name{dsrAssign}
\alias{dsrAssign}
\title{Load data in the remote session(s)}
\usage{
dsrAssign(
  symbol,
  table_or_sql,
  db_connection,
  async = TRUE,
  datasources = NULL
)
}
\arguments{
\item{symbol}{a character, the name of the data frame}

\item{table_or_sql}{a character, either the name of a table in the remote database or a valid SQL query}

\item{db_connection}{a character, the name of the SQLFlexClient object. This object has to exist already in the remote
session(s) and is created with the function datashield.assign.resource.}

\item{async}{same as in datashield.assign}

\item{datasources}{same as in datashield.assign}
}
\description{
Assign data to a data frame from a resource of the type SQL Query
}
