% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsShowColumns.R
\name{dsrShowColumns}
\alias{dsrShowColumns}
\title{Describe a table}
\usage{
dsrShowColumns(table_name, db_connection, async = TRUE, datasources = NULL)
}
\arguments{
\item{table_name}{a character, the name of the table (or view) as shown in the output of dsrShowTables(qualified with the schema name if the latter is not 'public', ex: 'information_schema.views')}

\item{db_connection}{a character, the name of the SQLFlexClient object. This object has to exist already in the remote
session(s) and is created with the function datashield.assign.resource}

\item{async}{same as in datashield.assign}

\item{datasources}{same as in datashield.assign}
}
\value{
a data frame containing the column names and data types
}
\description{
List the columns and data types of a table from a resource of the type SQL Query
}
